🐛 Roll-back switch to pyproject.toml
Apparently that makes this plugin no longer find
the OctoPrint installation during its own setup
and causes the install to fail.

That is very much unexpected and will require
some research as to why, and how it can be
prevented, so that in the future this change
can be applied again, and to the plugin template.

Closes #6